# bff-server/swagger.yaml
openapi: 3.0.3
info:
  title: Moodle BFF API
  description: Backend For Frontend API for Moodle integration
  version: 1.0.0
  contact:
    name: Moodle BFF API Support
servers:
  - url: https://52.194.117.196
    description: Production server (HTTPS via nginx)
  - url: http://52.194.117.196
    description: Development server (HTTP, redirects to HTTPS)
  - url: http://localhost:3001
    description: Local development (direct access)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Cognito JWT認証を使用します。

        1. Cognitoでログインして IDToken を取得
        2. "Authorize" ボタンをクリック
        3. IDToken を入力（"Bearer" プレフィックスは不要）

        トークン取得方法:
        ```bash
        # AWS CLIで取得
        aws cognito-idp initiate-auth \
          --auth-flow USER_PASSWORD_AUTH \
          --client-id 23jacbr6nk4baiftjueddmr4kb \
          --auth-parameters USERNAME=your-email,PASSWORD=your-password \
          --region ap-northeast-1

        # またはスクリプトを使用
        ./get-token.sh your-email your-password
        node test-auth.js your-email your-password
        ```
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        detail:
          type: string
...
    Badge:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        imageurl:
          type: string
...
    Roadmap:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        category:
          type: string
        difficulty:
          type: string
        description:
          type: string
...
paths:
...
  /api/webcoach/profile/{userid}:
    get:
      tags:
        - WebCoach
      summary: Get user profile
      description: Get WebCoach profile information for a user from the API server
      security:
        - bearerAuth: []
      parameters:
        - name: userid
          in: path
          required: true
          schema:
            type: integer
          description: User ID
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - can only access own profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Failed to get profile from API server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - WebCoach
      summary: Update user profile
      description: |
        Update WebCoach profile information. This endpoint forwards the request to the API server.

        All fields are optional. Only the fields you want to update need to be included in the request.
      security:
        - bearerAuth: []
      parameters:
        - name: userid
          in: path
          required: true
          schema:
            type: integer
          description: User ID (must match the authenticated user's ID)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileUpdate'
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - can only update own profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Failed to update profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/webcoach/updateprofile/{userid}:
    post:
      tags:
        - WebCoach
      summary: Update user profile (alias)
      description: |
        Update WebCoach profile information. This is an alias for POST /api/webcoach/profile/{userid}.

        All fields are optional. Only the fields you want to update need to be included in the request.
      security:
        - bearerAuth: []
      parameters:
        - name: userid
          in: path
          required: true
          schema:
            type: integer
          description: User ID (must match the authenticated user's ID)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileUpdate'
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - can only update own profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Failed to update profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/webcoach/resumecourse/{userid}:
    get:
      tags:
        - WebCoach
      summary: Get resume courses
      description: Get recommended courses to resume for a user with progress calculated from activity completion status
      security:
        - bearerAuth: []
      parameters:
        - name: userid
          in: path
          required: true
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 5
      responses:
        '200':
          description: List of courses to resume with progress
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ResumeCourse'
    post:
      tags:
        - WebCoach
      summary: Update resume course data
      description: Update or create resume course information for a user
      security:
        - bearerAuth: []
      parameters:
        - name: userid
          in: path
          required: true
          schema:
            type: integer
          description: User ID (must match the authenticated user's ID)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - courseid
                - progress_percent
              properties:
                courseid:
                  type: integer
                  description: Moodle course ID
                  example: 1
                progress_percent:
                  type: integer
                  minimum: 0
                  maximum: 100
                  description: Course completion percentage (0-100)
                  example: 0
      responses:
        '200':
          description: Resume course data updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResumeCourse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - can only update own data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Failed to update resume course data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/webcoach/recomendbadge/{userid}:
    get:
      tags:
        - WebCoach
      summary: Get recommended badges
      description: Get recommended badges for a user
      security:
        - bearerAuth: []
      parameters:
        - name: userid
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: List of recommended badges
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Badge'
...
  /api/webcoach/ai-applications:
    get:
      tags:
        - WebCoach
      summary: Get AI applications
      description: Get a list of AI applications with optional filtering
      security:
        - bearerAuth: []
      parameters:
        - name: category
          in: query
          schema:
            type: string
          description: Filter by category
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
          description: Maximum number of results to return
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
          description: Number of results to skip (for pagination)
      responses:
        '200':
          description: List of AI applications
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    name:
                      type: string
                    description:
                      type: string
                    category:
                      type: string
                    url:
                      type: string
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Failed to get AI applications
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/webcoach/roadmaps:
    get:
      tags:
        - WebCoach
      summary: Get roadmaps
      description: Get list of learning roadmaps
      security:
        - bearerAuth: []
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: difficulty
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of roadmaps
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Roadmap'
  /api/webcoach/roadmap/{roadmapid}:
    get:
      tags:
        - WebCoach
      summary: Get roadmap details
      description: Get detailed information about a specific roadmap
      security:
        - bearerAuth: []
      parameters:
        - name: roadmapid
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Roadmap details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Roadmap'
  /api/webcoach/ai:
    post:
      tags:
        - WebCoach
      summary: AI chat
      description: Interact with the WEBCOACH AI assistant
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AIRequest'
      responses:
        '200':
          description: AI response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIResponse'
  /api/webcoach/updatedb:
    post:
      tags:
        - WebCoach
      summary: Update database
      description: Bulk update WebCoach custom tables from CSV data
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDBRequest'
      responses:
        '200':
          description: Database updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  recordsProcessed:
                    type: integer
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/webcoach/faiss/stats:
    get:
      tags:
        - WebCoach
      summary: Get FAISS statistics
      description: |
        FAISSベクトルDBの統計情報を取得します。

        Returns:
        - 総ドキュメント数
        - 総ベクトル数
        - ベクトルの次元数
        - 使用している埋め込みモデル
        - S3バケット名とプレフィックス
      security:
        - bearerAuth: []
      responses:
        '200':
          description: FAISS statistics retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_documents:
                    type: integer
                    description: 総ドキュメント数
                  total_vectors:
                    type: integer
                    description: 総ベクトル数
                  dimension:
                    type: integer
                    description: ベクトルの次元数
                  embedding_model:
                    type: string
                    description: 使用している埋め込みモデル
                  s3_bucket:
                    type: string
                    description: S3バケット名
                  s3_prefix:
                    type: string
                    description: S3プレフィックス
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Failed to get FAISS stats
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/webcoach/faiss/reload:
    post:
      tags:
        - WebCoach
      summary: Reload FAISS index from S3
      description: |
        S3から最新のFAISSインデックスをリロードします。

        新しいデータがS3にアップロードされた後に実行してください。
        通常は取り込み処理後に自動実行されますが、緊急時やトラブルシューティング用に手動実行も可能です。
      security:
        - bearerAuth: []
      responses:
        '200':
          description: FAISS index reloaded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  stats:
                    type: object
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Failed to reload FAISS index
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...